fixed viewtopic page checking for an unset user_id variable bug

+++ viewtopic.php
@@ -63,7 +63,7 @@ $content .= '</dl>';
 // Post details
 $content .= '<div>';
 // Delete button
-if ($_SESSION['user_id'] === $result['user_id']) {
+if (isset($_SESSION['user_id']) && $_SESSION['user_id'] === $result['user_id']) {
     $content .= '<a class="button" href="deletetopic.php?id=' . $result['id'] . '">Delete</a>';
 }
 $content .= htmlspecialchars($result['content']);
